# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CBS Swagger backend server.",
        "title": "CBS Backend Swagger server",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/googlesearch/geosearch/:lat/:lng": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "GET request for closest address based on location latlon",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "googleapi"
                ],
                "summary": "Gets closest address based on location",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "userID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Latitude",
                        "name": "lat",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Longitude",
                        "name": "lng",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/maps.PlacesSearchResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    }
                }
            }
        },
        "/googlesearch/textsearch/:query": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "GET request for location strings based on search query",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "googleapi"
                ],
                "summary": "Gets locations from text query",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Query",
                        "name": "query",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "userID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/maps.PlacesSearchResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    }
                }
            }
        },
        "/googlesearch/textsearch/:query/:lat/:lng": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "GET request for location strings based on search query and latlon of user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "googleapi"
                ],
                "summary": "Gets prioritized locations from text query and position",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "userID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Query",
                        "name": "query",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Latitude",
                        "name": "lat",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Longitude",
                        "name": "lng",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/maps.PlacesSearchResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    }
                }
            }
        },
        "/locations/GetLatestLocationsPeriods/{secureId}": {
            "get": {
                "description": "Get latest location periods from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location"
                ],
                "summary": "Get latest location periods",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Secure ID",
                        "name": "secureId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.DeltaDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    }
                }
            }
        },
        "/locations/getLatest/{secureId}": {
            "get": {
                "description": "Get latest location from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location"
                ],
                "summary": "Get latest location",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Secure ID",
                        "name": "secureId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Location"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.DeltaDTO": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "integer"
                },
                "delta": {
                    "type": "integer"
                }
            }
        },
...
        "maps.AddressGeometry": {
            "type": "object",
            "properties": {
                "bounds": {
                    "$ref": "#/definitions/maps.LatLngBounds"
                },
                "location": {
                    "$ref": "#/definitions/maps.LatLng"
                },
                "location_type": {
                    "type": "string"
                },
                "types": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "viewport": {
                    "$ref": "#/definitions/maps.LatLngBounds"
                }
            }
        },
        "maps.LatLng": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                }
            }
        },
        "maps.LatLngBounds": {
            "type": "object",
            "properties": {
                "northeast": {
                    "$ref": "#/definitions/maps.LatLng"
                },
                "southwest": {
                    "$ref": "#/definitions/maps.LatLng"
                }
            }
        },
        "maps.OpeningHours": {
            "type": "object",
            "properties": {
                "open_now": {
                    "description": "OpenNow is a boolean value indicating if the place is open at the current time.\nPlease note, this field will be null if it isn't present in the response.",
                    "type": "boolean"
                },
                "periods": {
                    "description": "Periods is an array of opening periods covering seven days, starting from Sunday,\nin chronological order.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maps.OpeningHoursPeriod"
                    }
                },
                "permanently_closed": {
                    "description": "PermanentlyClosed indicates that the place has permanently shut down. Please\nnote, this field will be null if it isn't present in the response.",
                    "type": "boolean"
                },
                "weekday_text": {
                    "description": "weekdayText is an array of seven strings representing the formatted opening hours\nfor each day of the week, for example \"Monday: 8:30 am – 5:30 pm\".",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "maps.OpeningHoursOpenClose": {
            "type": "object",
            "properties": {
                "day": {
                    "description": "Day is a number from 0–6, corresponding to the days of the week, starting on\nSunday. For example, 2 means Tuesday.",
                    "type": "integer"
                },
                "time": {
                    "description": "Time contains a time of day in 24-hour hhmm format. Values are in the range\n0000–2359. The time will be reported in the place’s time zone.",
                    "type": "string"
                }
            }
        },
        "maps.OpeningHoursPeriod": {
            "type": "object",
            "properties": {
                "close": {
                    "description": "Close is when the place closes.",
                    "$ref": "#/definitions/maps.OpeningHoursOpenClose"
                },
                "open": {
                    "description": "Open is when the place opens.",
                    "$ref": "#/definitions/maps.OpeningHoursOpenClose"
                }
            }
        },
        "maps.Photo": {
            "type": "object",
            "properties": {
                "height": {
                    "description": "Height is the maximum height of the image.",
                    "type": "integer"
                },
                "html_attributions": {
                    "description": "htmlAttributions contains any required attributions.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "photo_reference": {
                    "description": "PhotoReference is used to identify the photo when you perform a Photo request.",
                    "type": "string"
                },
                "width": {
                    "description": "Width is the maximum width of the image.",
                    "type": "integer"
                }
            }
        },
        "maps.PlacesSearchResult": {
            "type": "object",
            "properties": {
                "business_status": {
                    "description": "BusinessStatus is a string indicating the operational status of the\nplace, if it is a business.",
                    "type": "string"
                },
                "formatted_address": {
                    "description": "FormattedAddress is the human-readable address of this place",
                    "type": "string"
                },
                "geometry": {
                    "description": "Geometry contains geometry information about the result, generally including the\nlocation (geocode) of the place and (optionally) the viewport identifying its\ngeneral area of coverage.",
                    "$ref": "#/definitions/maps.AddressGeometry"
                },
                "icon": {
                    "description": "Icon contains the URL of a recommended icon which may be displayed to the user\nwhen indicating this result.",
                    "type": "string"
                },
                "id": {
                    "description": "ID is an identifier.",
                    "type": "string"
                },
                "name": {
                    "description": "Name contains the human-readable name for the returned result. For establishment\nresults, this is usually the business name.",
                    "type": "string"
                },
                "opening_hours": {
                    "description": "OpeningHours may contain whether the place is open now or not.",
                    "$ref": "#/definitions/maps.OpeningHours"
                },
                "permanently_closed": {
                    "description": "PermanentlyClosed is a boolean flag indicating whether the place has permanently\nshut down.",
                    "type": "boolean"
                },
                "photos": {
                    "description": "Photos is an array of photo objects, each containing a reference to an image.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maps.Photo"
                    }
                },
                "place_id": {
                    "description": "PlaceID is a textual identifier that uniquely identifies a place.",
                    "type": "string"
                },
                "price_level": {
                    "description": "PriceLevel is the price level of the place, on a scale of 0 to 4.",
                    "type": "integer"
                },
                "rating": {
                    "description": "Rating contains the place's rating, from 1.0 to 5.0, based on aggregated user\nreviews.",
                    "type": "number"
                },
                "types": {
                    "description": "Types contains an array of feature types describing the given result.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_ratings_total": {
                    "description": "UserRatingsTotal contains total number of the place's ratings",
                    "type": "integer"
                },
                "vicinity": {
                    "description": "Vicinity contains a feature name of a nearby location.",
                    "type": "string"
                }
            }
        },
...
        "models.ErrorLog": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "deviceID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Location": {
            "type": "object",
            "properties": {
                "accuracy": {
                    "type": "number"
                },
                "altitude": {
                    "type": "number"
                },
                "bearing": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "integer"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "provider": {
                    "type": "string"
                },
                "sensorType": {
                    "type": "string"
                },
                "speed": {
                    "type": "number"
                },
                "trackerID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuthorization": {
            "type": "basic"
        }
    }
}